Add parallelization to renderers
RayTraceRenderer now only works in parallel since the gains are obvious
and the code is simple.
RasterizerRenderer has it's single threaded and parallel implementation
because the parallel implementation is a bit more involved and the gains
are not as big. This way they can be compared if necessary.

Some other minor refactoring.